+++ neo/tools/common/OpenFileDialog.h
@@ -96,7 +96,7 @@ ID_INLINE void rvOpenFileDialog::SetOKTitle ( const char* title )
 
 ID_INLINE void rvOpenFileDialog::SetInitialPath ( const char* path )
 {
-	if ( !idStr::Cmpn( mLookin, path, strlen( path ) ) )
+	if ( !idStr::Cmpn( mLookin, path, (int)strlen( path ) ) )
 	{
 		return;
 	}
