+++ neo/tools/materialeditor/ConsoleView.cpp
@@ -281,7 +281,7 @@ const char *ConsoleView::TranslateString(const char *buf) {
 	int			i, l;
 	char		*out;
 
-	l = strlen(buf);
+	l = (int)strlen(buf);
 	out = buf2;
 	for (i = 0; i < l; i++) {
 		if (buf[i] == '\n') {
